+++ lib/tie_fp.dart
@@ -1,3 +1,4 @@
 library tie_fp;
 
+part 'src/extensions.dart';
 part 'src/result.dart';