+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Add `toResult()` extension on `Future<T>`
+- Add `stackTrace` option on `Failure`
+
 ## 0.0.2
 
 - Minimal changes.